requirements.txt
django
django-guardian
channels_redis
psycopg2
pillow
selenium